Introduce boresight cal to counter Coupling atrifacts
In this commit, I have introduced bore-sight cal feature to counter the random phase cuased due to coupling from neighbouring Txs and Rxs. The following are the changes made in this commit.
1. Added a flag to enable or disable bore-sight cal
2. Defined the cal target at about 5m (closest to 5m which falls on a bin to eliminate bin quantization artifacts) and 0 velocity
3. Generating the signal for the cal target just like other targets and summing the actual signal with the cal target signal.
4. Generating the MIMO coefficients for the cal target and using this as the cal for the other objects.

However, the cal doesn't seem to be helping. It doesnt show any improvement in performance wil cal being applied. I need to debug this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>