Introduced the true impact of bore-sight cal in DDMA scheme
In this commit, I have introduced the true impact of the bore-sight caliberation routine for angle estimation. Previously also, I had introduced a bore-sight target to enable angle caliberation, but somehow, the effect of bore-sight cal was not seen. In other words, I was getting the same performance of the SLLs(in the angle spectrum) with/without boresight caliberation. In both cases, the SLLs were good even though a random phase was being introduced in the IC coupling as well as antenna coupling. This was puzzling since this same behaviour with/without cal is not expected. Now I have understood the reason why we observed good SLLs with/without bore-sight caliberation.

When the random phase was being introduced in the coupling matrix, I was forcing the diagonal elements to have 0 phase or same phase. This means that, predominantly, there is no random phase across the 4 transmitting Txs and all the Txs were more or less having 0 phase delta innately. There are however, some small random phases coupling into the reference Tx from other Txs. But because the coupling is 20-30 dB lower in power, hence the impact of these random coupling phases is also quite minimal. So, because of forcing the diagonal elements in the coupling matrix to have 0 phase, I had unknowingly removed the effect of bore-sight caliberation. Hence, I was getting the same good performance with/without cal. Now I have fixed this. I'm now introducing random phase across the Txs. This random phase variation across Txs is captured when the boresightCal flag is enabled and compensated for by sampling the ULA of the boresight target and treating it as cal. The other change I have introduced in this commit is I have disabled the phaseDemodMethod flag. I noticed that the MIMO coefficients for both the cal target and the other random targets needs to be genearted in the same fashion i.e. if phaseDemodMethod flag is enabled, both the cal target as well as the other taregts should have the MIMO coefficients generated through the demodulation procedure. Similarly, if the phaseDemodMethod is disabled. But if we have the cal target's MIMO coefficients generated using Dopler modulation sampling and the other targets' MIMO coefficients generated using base band demodulation, then somehow I'm seeing ambiguous behaviour and the caliberation is not working correctly. In fact in such a scenario, I get better performance without bore-sight cal and poor performance with bore-sight cal. I have not understood this!

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>