Cal works for phaseDemodMethod = 0 as well as 1
In this commit, I have found out the reason why the bore-sight cal was taking effect only for the case when phaseDemodMethod=0 and wasn't taking effect when phaseDemodMethod=1.
The reason is as follows:
I was adding the random phase to the Tx phase code signal namely "phaseCodesToBeApplied_rad" and in the phaseDemodMethod=1 mode, I was demodulating the chirp samples with the phasor correspoding to "phaseCodesToBeApplied_rad". Due to this, whatever random phase was introduced by the variable "phaseCodesToBeApplied_rad" was being removed at the receiver since I was demodulating using the same variable. Hence, I was not observing any impact on the angle SLLs with/without cal. Mathematically, I was introducing random phase to the phase code signal as exp(1j*(p*deltaPhi*m + thetaP)) where thetaP is the random phase dor each Tx. In the demodulation mehtods, I was again demodulating the chirp samples with exp(1j*(p*deltaPhi*m + thetaP)) hence in the process of demodulation, the random phase was also removed and hence no imapct of cal.
But I was observing the impact when "phaseDemodMethod" was set to 0 because, in this method, there is no demodulation involved rather sample the Doppler spectrum(or evaluate using DFT) at the modulated bin locations. When we are evaluating the spectrum at the modulated bin locations, the random phase introduced still holds and hence we see the impact of bore-sight cal.

Now I have fixed this issue in the phaseDemodMethod=1 mode by assigning a separate variable for the random phase and not as part of the phaseCodesToBeApplied_rad phasor. That way, the phasor corresponding to the Tx phase codes is clean. At the receiver end, I demodulate with this clean phasor corresponding to the Tx phase codes and hence the arbitrary phase introduced still remains. Now with this change, I see the impact of bore-sight cal both in phaseDemodMethod=0 and phaseDemodMethod=1 modes.

This was a nice challenging problem that took me almost 1 week to understand and debug.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>