Compute Zeros of a sinusoid transfer function in an optimized manner
In this commit, I have developed a simple method of computing the zeros of a sinusoid transfer function. Ideally, in order to find the zeros of an FIR transfer function, we need to compute the Z transform of the transfer function/impulse response and then solve for the roots of the polynomial in Z inverse to obtain the roots/zeros of the transfer function. However, if the signal/impulse response/transfer function is a sinusoid, we don't need to solve for the roots of a polynomial(which is a compute heavy operation). Instead, we can find them in a simple manner. We know that if we have a sinusoidal signal of digital frequency w and of N samples, if we perform an FFT, the zeros are separated by 2*pi/N. These zeros/nulls will be visible with a simple N point FFT only for those sinusoidal signals whose digital frequency is an integer multiple of 2*pi/N. For sinusoidal signals whose digital frequency is not an integer multiple of 2*pi/N, a simple N point FFT will not reveal the nulls or zeros. To visualize the zeros/nulls, we need to perform an over sampled FFT. This is becuase, if the digital frequency doesnt fall on a bin, then the zeros/nulls which occur at a separation of 2*pi/N from the signal peak are also not evaluted when a simple N point FFT is performed. However, when an oversampled FFT is performed, we are essentailly evaluating the spectrum on a much finer grid and hence we will ineveitable end up evaulating the spectrum at a frequency point which is 2*pi/N away from the signal peak as well. Hence we will start observing the nulls/zeros. Since this is establised, to compute the zeros for the sinusoidal transfer function of N samples, all we need to do is to add integer multiples of 2*pi/N to the signal digital frequency in order to obatin the zeros. If the signal has a length of N, then there will be N-1 zeros. So, if the sinusoidal signal has a digital frequency w, then the zeros will be w + 2*pi/N * n, where n goes from (1,N-1). If n is 0 or N, then we land on the signal frequency and hence these points are avoided. Prevously, to compute the zeros of even a sinusoidal transfer function, I was calling the tf2zpk function from the scipy.signal library. This function was essentially taking the Z transform of the signal and then solving for the roots of the polynomial. Now, I have understood the structure of a sinusoidal sigal and hence compute the zeros in a trivial manner. This reduces compute significant. This is a new trick I have realized!

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>