Spatially Variant Apodization
In this commit, I have implemented a technique called Spatially Variant Apodization (SVA). SVA is a like for like replacement for a windowed FFT. A typical windowed FFT is characterized by main lobe width and side lobe level. Rectangular window has very good main lobe width but poor SLLs (13 dBc). Hanning window has poor main lobe width but good side SLLs (31 dBc). SVA is a technique which combines the best of both these windows. It gives the main lobe width of a rectangular window while giving the SLL of a Hanning window. Thus it gives the best of both worlds. The working principle of SVA is as follows:

The SVA attemps to select a different window function for each frequency. Let us understand this more closely. A raised cosine window is a generic window which is controlled by an alpha parameter. The functional form of the window is 1-2*alpha*cos(2*pi/N*n), where 'alpha' lies in [0,0.5], 'N' is the number of samples and 'n' varies from 0 to N-1. When 'alpha' = 0, the raised cosine window becomes a rectangular window. When 'alpha' = 0.5, the window becomes a hanning window. By varying alpha from 0 to 0.5 in fine steps, we can create multiple variants of the raised cosine window. A brute force version of the SVA is as follows. Subject the signal to each of these windows and perform an FFT on each of the windowed signals. This results in several FFTs of the signal each with a different window. Now, compute the psd/magnitude square of each of these FFTs. To obtain the final psd/spectrum, take minimum along the window axis. This results in a spectrum which has minimum energy at each frequency point across all window function. In essence, we are selecting a window function which generates the minimum energy at each frequency point. Since, we are selecting munimum energy at each frequency point (across all windows), this will result in minimum SLLs as well as best main lobe width. But this brute force method requires heavy compute since we need to generate several windows and multiply the signal also with each of these windows and compute FFT on each of these signals. Finally we also need to take a minimum for each of the frequency point. This is serioulsy heavy compute as well as memory. This can be circumvented by implementing the optimized version of the SVA. For this, we need to take a closer look at the structure of the raised cosine window. The functional form of the window is w[n] = 1-2*alpha*cos(2*pi/N*n). If we assume the signal of interest is x[n], then the windowed signal is y[n] = x[n]w[n]. Now multiplication in time domain is equivalent to convolution in the frequency domain. So if we were to analyse the windowed signal in the Fourier domain, we get Y(w) = X(w) * W(w), where * denotes convolution operation. Now W(w) = 2pi * ( delta(w) - alpha*[delta(w-2*pi/N) + delta(w+2*pi/N)]). So,  Y(w) = 2pi * (X(w) - alpha*[X(w-2pi/N) + X(w+2pi/N)]). Now, the optimum value of alpha(at each frequency) is chosen by minimizing the magnitude squared of Y(w) over all alpha lying between 0 to 0.5. This is a constrained optimization problem. Solving this, we obtain a closed form expression for alpha at each frequency point omega. Hence alpha is a function of the omega chosen. The other details of the derivation are available in the paper link below:
https://saigunaranjan.atlassian.net/wiki/spaces/RM/pages/22249477/Spatially+Variant+Apodization

After obtaining a closed form expression for alpha and substituting in the expression for Y(w), we obatin the spectrum for SVA. This method of computing the SVA spectrum is computationally light. As evident from the closed form expression for alpha, the only major compute (apart from the FFT computation of x[n]) is an N point division for an N point oversampled FFT. So there are N complex divisions corresponding to each of the N frequency points.

I have implemented both the brute force method as well as the optimized method of SVA and have generated the results. Both thse methods seem to be closely matching. The results are on expected lines with the SVA offering the best SLLS as well as main lobe width.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>