Impact of Doppler residual on angle accuracy and SLLs
In this commit, I have modelled the impact of Doppler residual phase on the ULA azimuth spectrum. More specifically, I will further analyze the impact of an inaccurate Doppler estimate on the angle accuracy and SLLs. Since the Doppler is used to correct/compensate for the phase jump introduced due to phase change across chirps in a TDM system, an inaccurate Doppler estimate results in an incorrect Doppler phase correction thus deteriorating the angle accuracy and SLLs. I will further test this using oversampled Doppler FFT of different lengths like 256, 512, 1024 and so on. This will reveal what accuracy is required in the Doppler estimate for a good angle accuracy and SLLs.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>