Fix bug in generation and application of Dopp correction
In this commit, I have fixed a bug in the generation and application of the Doppler correction phase term for the MIMO syntesis. This debug helped me to thoroughly understand the Doppler phase correction. The point to understand is that a true Doppler bin doesn't have to be an integer bin. By doing an FFT and picking a peak, we can only sample integer bins. So, this fractional error i.e true Doppler bin (which is a float) - detected Doppler bin (which is necessarily an integer) leads to an uncompensated phase error in the MIMO array leading to angle error and SLLs. If the true Doppler bin lies exactly on a integer grid(with no fractional), then the Doppler correction phase term becomes 1 and hence no doppler correction phase is required. Eg: If Vtrue/vRes is an integer, then the DCM becomes unity (by chance!). But we have no control on the true velcoity of the target and hence it might or might not fall on a bin. So we need to get as close to the true Doppler bin (float) as possible. This is achieved through an oversampled FFT evaluation. Let us undersatnd this more closely. Consider a case where we have 64 ramps per Tx and 4 Txs. So the velres = Fc/64 * lamda/2. The DCM becomes exp (1j*2*pi*k/64*n*64), where n is the Tx number and n varies from 0 to Ntx-1; k is Doppler bin = v/vRes. Now, if k is exactly an integer and detection also picks up the same integer when 64 point Doppler FFT is performed on the 64 chirp samples, then, k becomes an integer and the DCM becomes unity (64 cancels out in numerator and denominator, k and n take integer values). In this case, luckily/coincidentally, no DCM is required while synthesizing the ULA. But consider a case when k is not an integer. Now performing a Doppler FFT and picking the peak gives us a qunatized version of the k i.e we get int(k). As seen earlier, if k is an integer (on the 64 scale), then the DCM terms becomes unity and this essentially means we are not applying any DCM. But because of the error between the true Doppler bin and the quantized Doppler bin(which we call as the Doppler fractional error), there is a Doppler residual phase on the MIMO which is un corrected. This leads to the ULA phase have phase jumps leading to angle error and poor SLLs in the angle spectrum. To reduce this Doppler fractional error, we need to come as close to the tru Dopplr frequency as possible. This is achieved by using a oversampled FFT to obatin the Doppler bin (on the oversampled scale). Since an oversampled FFT gives a more accurate estimate of the true Doppler frequency, we can get closer to the true Doppler frequency. So, in the DCM phase term, the denominator is no longer 64 but may be 512 or 1024 depending on what point Doppler FFT we are performing. So now, the numerator (64) and denominator do not cancel out anymore and also the Doppler bin 'k' is no more on the 64 scale but it is on the oversampled FFT scale. The greater the oversampling factor (for the FFT computation), the closer we can get to the true Doppler frequency and smaller the DCM error. But upto what point do we need to go? In toher words whats is the oversampling FFT factor upto which the DCM error is tolerable. This we can get through a Monte Carlo siumation run. I will perform the MC with different overampling factors in the next commit.
Some of the other minor corrections made in this commit are:
1. Changed the number of chirps from 256 tp 512. But we use only 256 chirps for our analysis. So we have 4 Txs each transmitting 64 chirps.
2. Introduced a Doppler OSR factor to control the number of points for the oversampled Doppler FFT.
3. There was a bug in the construction of Doppler phase term and the doppler phase across Tx term. Since I'm analysing the DCM on a radar system model where in we have equal number of ramps per Tx, the denominator in the Doppler phase term should be numRampsPerTx. Previosuly, it was total number of ramps in detection segment. This was wrong. Now, I have corrected it.
4. Detecting the Doppler bin on an oversampled scale for each of the Monte carlo runs.
5. Fixed a minor bug in plotting the unwrapped phase of the ULA. I was plotting the unwrapped phase of the ULA post multiplying the MIMO phasor with a window. Becuase of this I was observing a phase jump at the last MIMO sample. I was puzzled by this. It took me a long time to root cause the issue to the window application on the ULA phasor. Since the window was being applied on the MIMO phasor before unwrapping the phase, the last sample of the window (which is close to 0), was destroying the last element of the ULA phasor signal. Thus when I was unwrapping the phase after this, I was observing a phase jump. Now, I first unwrap the phase of the ULA phasor and store it in a variable then I apply windowing on the MIMO phasor to perform the angle FFT subsequently.
6. The detected Doppler bin for the DCM correction is now on an oversampled scale.
7. Logging the angle error and SLLs which can be used later for the MC run.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>