Provide a test case script to compare 2 flavours of MUSIC
In this commit, I have added a test case script to compare the
performance of a sliding window based MUSIC and a snapshot based MUSIC
algorithms. We can clearly see the resolution performance improvement
when multiple snapshots are actually available. Also, the spatial noise smooting step i.e. (R + JR*J)*0.5, helps to further improve the resolution With an 8 channel MUSIC, the ideal rayleigh resolution limit is about 14 degrees. But with the snapshot based MUSIC and the noise smooting operation , we can hit upto 1.4 degrees which is a staggering 1/10 of the rayleigh limit.
I was aware that for MUSIC enhanced performance, we neeed multiple
snapshots, but since in reality, especially for RADAR, where the scene
is constantly changing, multiple snapshots are ideally not available so
I did not bother to code the multiple snapshot based MUSIC. Now I see
the actual performance improvement with multiple snapshots.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>