Implement pair wise azimuth/elevation association
In this commit, I have implemented a paper which performs a pair wise association of the azimuth and elevation peaks for an L shaped array.
The implementation is based on the paper 'Pair-Matching Method for Estimating 2-D Angle of Arrival With a Cross-Correlation Matrix' by Kikuchi et al. and published in IEEE ANTENNAS AND WIRELESS PROPAGATION LETTERS. The paper is very implementation friendly and I have implemented the paper as is. The paper assumes a slighly different phase model for the azimuth and elevation arrays. Typically,
x = cos(theta)
y = sin(theta) cos(phi)
z = sin(theta) sin(phi)
But in this papers, the model is as follows:
y = sin(phi)
z = sin(theta)
It is as if the point moves in azimuth first and then in elevation like the positioner movement.
This paper clearly mentions the algorithm as a step by step procedure which can easliy be followed.
In this commit, I have tested the algo with 3 targets in azimuth and 3 in elevation with 3 different SNRs for each of the 3 targets. I have also assumed that the azimuth and elevation arrays are identical and have same elemental spacing. In the subsequent commits, I will test with non-identical arrays which have different spacing for azimuth and elevation arrays.
Also, this paper assumes both the azimuth and elavtion arrays pick up the same number of peaks. It doesn't handle the case when there are different number of peaks for the azimuth and elevation arrays.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>