Perform SVA for a multidimensional signal across specified axis
In this commit, I have extended the Spatially Variant Apodization
algorithm to cater to a multidimensional signal and perform SVA across
any specified axis. Towards this, I have made use of the np.take
function which samples a multi dimensional signal using an array along a
specified dimension. To verify this, I have created a test case as well.
I create a 2D sinusoidal signal and perform SVA along both the axis with
axis as input. I then check the correctness of the SVA by comparing
against the FFT across the 2 dimensions.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>