Fix bug in computing SLLs
In this commit, I have identified and fixed a bug in estimating the SLLs
of a signal. I estimate the SLLs using the argrelextrema function which
picks the local maxima. The side lobe value is captured by picking the second largest local maxima value. Now, this function takes in 2 main arguments namely
1. comparator: This is to choose the condition for the local maxima. So
if we want the local maxima to be defined as the value should be strictly
greater, or greater than or equal to its either side neighbours.
2. order: The local maxima should be greater than or greater than or
equal to 1 neighbour on eother sides or 2 neighbours on either sides or
3 neighbours on either sides, so on. So the order parameter defines this
number of neighbourhood samples.
Now, I was setting the arguments as strictly greater and giving order as
2. But this can fail/lead to false SLL values especially when the over
sampling factor is small like 2. Then, it can so happen that the SLL
peak might fall in between 2 bins and this will result in flat SLL peak
and hence the argrelextrema function will not pick this peak and will
search for the next local maxima satisfying the strictly greater than
condtion and that too with 2 neighbours. To avoid this, I have set the
comparator to greater than or equal to and set the order to 1. That way,
this caters to both the flar local maxima SLL as well as low
oversampling factor which has local maxima for only +/- 1 neighbours.

Also, did some code clean up.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>