Analyse SLLs of SVA vs singal length and no noise
In this commit, I have madea few changes to better understand the side
lobe levels behaviour of the SVA method. Towards this, I have removed
the addition of noise to the signal, since I dont want noise to limit
the SLLs. It appears that the SLLS of the SVA are not fixed and infact dependenton the length of the signal. The SLLs seem to increase with increase in signal length. This is a new discovery and I need to verify it a bit more thoroughly. I think it makes sense since the other windows as fixed functions and hence the SLLs of fft with wondow functions like hanning, rectangular, etc are constant. Whereas, SVA is a signal dependent window and hence maybe the dependence on signal length. Also, I will analyse the mainlobe width as well.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>