Compare scalloping loss of SVA against other windows
Scalloping loss is the loss in power, when the signal frequency falls in
the middle of 2 bins. This loss in power of the peak spills to the
neighbouring bins. Scalloping loss is a spec prameter to characterize window functions. The scalloping loss for a rectangular window is 3.92 dB and for hanning window it is 1.42 dB. I'm also able to exactly reproduce these numbers. The scalloping loss for SVA is also 3.92 dB i.e same as that of rectangular window.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>