Compute Scaling Factor without computing numLeadingZeros
In this commit, I have modified the computation of the scaling/shift
factor. Previously, I was first finding the number of leading zeros as
numLeadingZeros = DEN_TOT_BITS - positionOfLeading1fromLSB and then
computed the scaling facor as scalingFactor = DEN_INT_BITS - numLeadingZeros. But this is redundant. The more effective way to compute the scaling facor is
scalingFactor = positionOfLeading1fromLSB - DEN_FRAC_BITS. This doesnt
require number of integer bits.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>