Change the bitwidth of num and denom to 8Q8
In this commit, I have changed the bit width of the numerator and
denominator to 8Q8 format i.e both num and den are 16 bit numbers and
hence the divider is a 16 bit divider. Now, I can afford to accomodate
bit growth that occurs due to multiplication of d and x_i and also the
final multiplication of x_i with (2-d*x_i). Now at this final stage, we
can drop the bits (instead of dropping at an earlier stage and lose
precision). Since I'm now working with smaller bitwidths, I can afford
to do this. But if I'm working with larger bitwidths, it will bloat to
48 and 64 bits which cannot be contained in python. Hence I had to drop
bits at the earlier stage.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>