Added a newton raphson divider bitmatched to RTL
In this commit, I have added another script for the Newton Raphson
divider which has been bitmatched to the RTL logic. In this script, I
have slighly modified the implementation of the fixed point Newton
raphson algorithm. Now, I dont explicitly use the scaling factor (for
bringing the denominator to range [0.5,1)) in the Newton raphson update
equation. I apply the scaling to the final output post the Newton
Raphson iteration. I will explain this logic in a more detailed manner
as a doc string documentation in the script in the subsequent commits.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>