RFFT bitwidth vs quantization floor
In this commit, I have created a script to analyze the impact of different bitwidths at the Range FFT output. For example how does the range FFT floor compare for a 8 bit vs 16 bit vs 24 bit RFFT output. We need the quantization floor to be well below the thermal floor so that the SNR is not limited, else the quantization floor due to the LSB will start limiting due to underflows. The smallest absolute value for RFFT output with B fractional bits is 2**-B (1 LSB value). Hence the power spectrum will have 2B fractional bits and hence the smallest power value that can be represented is 2**(2*-B). This sets the noise floor(in linear scale). On log scale it is 10*np.log10(2**(2*-B)). So, all power values less than 2**(-2B) will underflow to 0.

I will be examining the impact of this qunatization and design a system with the optimal bitwidth to cater to range FFT specs and parameters such as system gain, detection theshold, etc.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>