Analyse the impact of bitwidth on different SNRs
In this commit, I have introduced multiple targets in the Range FFT with
SNR dropping from 60 dB to -15 dB as we move away from 0 range bin. Now,
we study the impact of different bitwidths (post RFFT, DFFT, Angle FFT)
on different SNR targets. We then understand when are the low SNR
targets detected or missed or detected but with wrong reported SNRs. We
also plot the reported SNR(with quantization) vs True SNR and show what
bitwitdh is required for correct detection of -15 dB SNR taregts in
RFFT. -15 dB Range FFT SNR will become 18 dB (detection SNR) post
Doppler and Angle FFT with a system gain of 512 chirps and 4 Rxs.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>