Speed up sorting of a list
In this commit, I have added a new function to sort the list of nodes by
their value. The previous implementation was sub-optimal and had two
nested for loops and hence the order of complexity was O(n2) and hence I
have implemented the sorting of the list of nodes using inbuilt sorting
functions for lists. This implementation is much faster than the earlier
one.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>