Show that Phase shifter quant noise is a dBc noise
In this commit, I have added a script to show how the phase shifter quantization noise (DNL) appears as a phase noise and hence limits the SNR for strong targets. More specifically, I will show that the phase shifter quantization noise is a dBc noise(phase noise) and scales with the signal power beyond a point. For weaker SNR targets, the phase noise floor will be buried below the thermal floor (post Doppler FFT). When the SNR of the targets increases and reaches  ~58 dB (DNL floor for 512 ramps), the SNR will saturate.

The following are the changes made to the DDMA_singleDoppler.py script to perform this analysis. I have named the new script as DDMA_phaseshifter_noise_analysis.
1. Made sure that the Tx modulations fall exactly on integer Doppler bins by djusting the phase shifter values for 512 ramps.
2. Removed the windowing for both RFFT as well as DFFT to avoid windowing effect and biasing the signal power and noise floor while estimating the SNR of the target. By removing the windowing artifacts, we can get an accurate estimate of noise floor, signal power and hence SNR which macthes the expected value. We can still apply windowing and estimate the SNR but we need to be careful while computing the signal power and noise floors. We need to adjust for the windowing arrtifacts for signal power and noise floor accordingly. There are closed form expressions for this. Fow now, I have avoided the windowing artifacts all together.
3. Since I will npt be computing the angle, I have removed the Tx contribution to the singal phase for the angle. Even the Rxs cantribution can be removed, but since I wanted some averaging effect for the Doppler spectrum, I have used Rx signal.
4. Made the vecolity of tghe taregt to 0 so that the base modulation falls exactly on 0 Doppler bins and the subsequent Tx modulations also fall on integer Doppler bins. This will avoid any spreading of the signal across bins and hence the comutation for singal power will be simple and accurate.
Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>